CAMEL-15584: Switch UUID generator to new fast vanilla implementation and move old into classic generator implementation.
