+++ components/camel-cdi/src/test/java/org/apache/camel/cdi/mock/DummyRestConsumerFactory.java
@@ -22,10 +22,10 @@ import org.apache.camel.CamelContext;
 import org.apache.camel.Consumer;
 import org.apache.camel.Processor;
 import org.apache.camel.component.seda.SedaEndpoint;
-import org.apache.camel.impl.engine.DefaultUuidGenerator;
 import org.apache.camel.spi.RestApiConsumerFactory;
 import org.apache.camel.spi.RestConfiguration;
 import org.apache.camel.spi.RestConsumerFactory;
+import org.apache.camel.support.ClassicUuidGenerator;
 
 public class DummyRestConsumerFactory implements RestConsumerFactory, RestApiConsumerFactory {
 
@@ -39,9 +39,9 @@ public class DummyRestConsumerFactory implements RestConsumerFactory, RestApiCon
         // just use a seda endpoint for testing purpose
         String id;
         if (uriTemplate != null) {
-            id = DefaultUuidGenerator.generateSanitizedId(basePath + uriTemplate);
+            id = ClassicUuidGenerator.generateSanitizedId(basePath + uriTemplate);
         } else {
-            id = DefaultUuidGenerator.generateSanitizedId(basePath);
+            id = ClassicUuidGenerator.generateSanitizedId(basePath);
         }
         // remove leading dash as we add that ourselves
         if (id.startsWith("-")) {
@@ -59,7 +59,7 @@ public class DummyRestConsumerFactory implements RestConsumerFactory, RestApiCon
             Map<String, Object> parameters)
             throws Exception {
         // just use a seda endpoint for testing purpose
-        String id = DefaultUuidGenerator.generateSanitizedId(contextPath);
+        String id = ClassicUuidGenerator.generateSanitizedId(contextPath);
         // remove leading dash as we add that ourselves
         if (id.startsWith("-")) {
             id = id.substring(1);
