+++ core/camel-base/src/main/java/org/apache/camel/impl/engine/SimpleCamelContext.java
@@ -76,6 +76,7 @@ import org.apache.camel.spi.UuidGenerator;
 import org.apache.camel.spi.ValidatorRegistry;
 import org.apache.camel.spi.XMLRoutesDefinitionLoader;
 import org.apache.camel.support.DefaultRegistry;
+import org.apache.camel.support.DefaultUuidGenerator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
