+++ core/camel-core/src/test/java/org/apache/camel/component/rest/DummyRestConsumerFactory.java
@@ -22,11 +22,11 @@ import org.apache.camel.CamelContext;
 import org.apache.camel.Consumer;
 import org.apache.camel.Processor;
 import org.apache.camel.component.seda.SedaEndpoint;
-import org.apache.camel.impl.engine.DefaultUuidGenerator;
 import org.apache.camel.spi.RestApiConsumerFactory;
 import org.apache.camel.spi.RestConfiguration;
 import org.apache.camel.spi.RestConsumerFactory;
 import org.apache.camel.support.CamelContextHelper;
+import org.apache.camel.support.ClassicUuidGenerator;
 
 public class DummyRestConsumerFactory implements RestConsumerFactory, RestApiConsumerFactory {
 
@@ -49,9 +49,9 @@ public class DummyRestConsumerFactory implements RestConsumerFactory, RestApiCon
         // just use a seda endpoint for testing purpose
         String id;
         if (uriTemplate != null) {
-            id = DefaultUuidGenerator.generateSanitizedId(basePath + uriTemplate);
+            id = ClassicUuidGenerator.generateSanitizedId(basePath + uriTemplate);
         } else {
-            id = DefaultUuidGenerator.generateSanitizedId(basePath);
+            id = ClassicUuidGenerator.generateSanitizedId(basePath);
         }
         // remove leading dash as we add that ourselves
         if (id.startsWith("-")) {
@@ -77,7 +77,7 @@ public class DummyRestConsumerFactory implements RestConsumerFactory, RestApiCon
             Map<String, Object> parameters)
             throws Exception {
         // just use a seda endpoint for testing purpose
-        String id = DefaultUuidGenerator.generateSanitizedId(contextPath);
+        String id = ClassicUuidGenerator.generateSanitizedId(contextPath);
         // remove leading dash as we add that ourselves
         if (id.startsWith("-")) {
             id = id.substring(1);
