+++ core/camel-support/src/main/java/org/apache/camel/support/DefaultUuidGenerator.java
@@ -22,13 +22,10 @@ import java.util.concurrent.atomic.AtomicLong;
 import org.apache.camel.spi.UuidGenerator;
 
 /**
- * An implementation of the {@link UuidGenerator} that uses a random generated seed and a counter which increments by
- * one. This generator is not unique per host or JVM.
- *
- * @deprecated this is now the default uuid generator in Camel 3.6 onwards.
+ * Default {@link UuidGenerator} optimized for Camel usage.
  */
-@Deprecated
-public class VanillaUuidGenerator implements UuidGenerator {
+public class DefaultUuidGenerator implements UuidGenerator {
+
     private static final char[] HEX_ARRAY = "0123456789ABCDEF".toCharArray();
 
     private final char[] seed
@@ -50,4 +47,5 @@ public class VanillaUuidGenerator implements UuidGenerator {
         }
         return new String(hexChars);
     }
+
 }
